feat: dont wrap gnot to wgnot in dashboard
