+++ packages/web/src/views/dashboard/components/activity-list/activity-list-table/activity-info/ActivityInfo.tsx
@@ -1,6 +1,8 @@
 import dayjs from "dayjs";
 import relativeTime from "dayjs/plugin/relativeTime";
+import { TFunction } from "i18next";
 import React, { ReactNode } from "react";
+import { useTranslation } from "react-i18next";
 
 import DateTimeTooltip from "@components/common/date-time-tooltip/DateTimeTooltip";
 import IconOpenLink from "@components/common/icons/IconOpenLink";
@@ -9,6 +11,13 @@ import {
   ACTIVITY_INFO,
   MOBILE_ACTIVITY_INFO,
 } from "@constants/skeleton.constant";
+import { useGnoscanUrl } from "@hooks/common/use-gnoscan-url";
+import { ActivityData } from "@repositories/activity/responses/activity-responses";
+import { DexEvent } from "@repositories/common";
+import {
+  formatOtherPrice,
+  formatPoolPairAmount,
+} from "@utils/new-number-utils";
 import { formatAddress } from "@utils/string-utils";
 
 import {
@@ -34,14 +43,24 @@ export interface Activity {
 }
 
 interface ActivityInfoProps {
-  item: Activity;
+  item: ActivityData;
   idx?: number;
   key?: number;
 }
 
 const ActivityInfo: React.FC<ActivityInfoProps> = ({ item }) => {
-  const { action, totalValue, tokenAmountOne, tokenAmountTwo, account, time } =
-    item;
+  const { t } = useTranslation();
+  const { getTxUrl } = useGnoscanUrl();
+
+  const {
+    action,
+    totalValue,
+    tokenAmountOne,
+    tokenAmountTwo,
+    account,
+    time,
+    explorerUrl,
+  } = formatActivity(item, t, getTxUrl(item.txHash));
 
   return (
     <TokenInfoWrapper>
@@ -51,7 +70,7 @@ const ActivityInfo: React.FC<ActivityInfoProps> = ({ item }) => {
             {action}
             <IconButton
               onClick={() => {
-                window.open(item.explorerUrl, "_blank");
+                window.open(explorerUrl, "_blank");
               }}
             >
               <IconOpenLink className="action-icon" />
@@ -92,8 +111,18 @@ const ActivityInfo: React.FC<ActivityInfoProps> = ({ item }) => {
 };
 
 export const MobileActivityInfo: React.FC<ActivityInfoProps> = ({ item }) => {
-  const { action, totalValue, tokenAmountOne, tokenAmountTwo, account, time } =
-    item;
+  const { t } = useTranslation();
+  const {getTxUrl} = useGnoscanUrl();
+
+  const {
+    action,
+    totalValue,
+    tokenAmountOne,
+    tokenAmountTwo,
+    account,
+    time,
+    explorerUrl,
+  } = formatActivity(item, t, getTxUrl(item.txHash));
 
   return (
     <MobileActivitiesWrapper>
@@ -106,7 +135,7 @@ export const MobileActivityInfo: React.FC<ActivityInfoProps> = ({ item }) => {
             {action}
             <IconButton
               onClick={() => {
-                window.open(item.explorerUrl, "_blank");
+                window.open(explorerUrl, "_blank");
               }}
             >
               <IconOpenLink className="action-icon" />
@@ -157,3 +186,103 @@ export const MobileActivityInfo: React.FC<ActivityInfoProps> = ({ item }) => {
   );
 };
 export default ActivityInfo;
+
+const formatActivity = (
+  res: ActivityData,
+  t: TFunction<"translation", undefined>,
+  explorerUrl: string,
+): Activity => {
+  const tokenASymbol = res.tokenA.symbol;
+  const tokenBSymbol = res.tokenB.symbol;
+  const shouldShowTokenAAmount =
+    res.actionType !== DexEvent.CLAIM_FEE ||
+    (!!res.tokenAAmount && !!Number(res.tokenAAmount));
+  const shouldShowTokenBAmount =
+    res.actionType !== DexEvent.CLAIM_FEE ||
+    (!!res.tokenBAmount && !!Number(res.tokenBAmount));
+
+  const actionText = (() => {
+    const action = (() => {
+      switch (res.actionType) {
+        case DexEvent.SWAP:
+          return t("business:onchainActi.action.swap");
+        case DexEvent.ADD:
+          return t("business:onchainActi.action.add");
+        case DexEvent.REMOVE:
+          return t("business:onchainActi.action.remove");
+        case DexEvent.DECREASE:
+          return t("business:onchainActi.action.decrease");
+        case DexEvent.INCREASE:
+          return t("business:onchainActi.action.increase");
+        case DexEvent.REPOSITION:
+          return t("business:onchainActi.action.reposition");
+        case DexEvent.CLAIM_FEE:
+          return t("business:onchainActi.action.claimFees");
+        case DexEvent.ADD_INCENTIVE:
+          return t("business:onchainActi.action.incentivize");
+        case DexEvent.STAKE:
+          return t("business:onchainActi.action.stake");
+        case DexEvent.UNSTAKE:
+          return t("business:onchainActi.action.unstake");
+        case DexEvent.CLAIM_REWARD:
+          return t("business:onchainActi.action.claimRewards");
+      }
+    })();
+
+    const tokenAText =
+      shouldShowTokenAAmount && tokenASymbol ? " " + tokenASymbol : "";
+    const tokenBText =
+      shouldShowTokenBAmount && tokenBSymbol ? " " + tokenBSymbol : "";
+
+    const relatedTokens = res.usedTokens || 0;
+
+    let conjunction = "";
+    if (relatedTokens === 2 && tokenAText && tokenBText) {
+      conjunction = ` ${
+        res.actionType === DexEvent.SWAP
+          ? t("common:conjunction.for")
+          : t("common:conjunction.and")
+      }`;
+    } else if (relatedTokens > 2) {
+      conjunction = ", ";
+    }
+
+    const tail = relatedTokens > 2 && ", ...";
+
+    return (
+      <span>
+        {action}
+        <span className="symbol-text">{tokenAText}</span>
+        {conjunction}
+        <span className="symbol-text">{tokenBText}</span>
+        {tail}
+      </span>
+    );
+  })();
+
+  const tokenAAmount =
+    tokenASymbol && shouldShowTokenAAmount
+      ? `${formatPoolPairAmount(res.tokenAAmount, {
+          decimals: res.tokenA.decimals,
+        })} ${res.tokenA.symbol}`
+      : "-";
+
+  const tokenBAmount =
+    tokenBSymbol && shouldShowTokenBAmount
+      ? `${formatPoolPairAmount(res.tokenBAmount, {
+          decimals: res.tokenB.decimals,
+        })} ${res.tokenB.symbol}`
+      : "-";
+
+  return {
+    action: actionText,
+    totalValue: formatOtherPrice(res.totalUsd, {
+      isKMB: false,
+    }),
+    tokenAmountOne: tokenAAmount,
+    tokenAmountTwo: tokenBAmount,
+    account: res.account,
+    time: res.time,
+    explorerUrl,
+  };
+};
